SchemaBuilder restructuring round. Jesus take the keyboard.
