+++ query-engine/core/src/schema_builder/input_types/input_fields.rs
@@ -7,7 +7,7 @@ pub(crate) fn filter_input_field(ctx: &mut BuilderContext, field: &ModelField) -
 }
 
 pub(crate) fn nested_create_input_field(ctx: &mut BuilderContext, field: &RelationFieldRef) -> InputField {
-    let create_types = create_input_objects::create_input_types(ctx, &field.related_model(), Some(field));
+    let create_types = create_objects::create_input_types(ctx, &field.related_model(), Some(field));
     let types: Vec<InputType> = create_types
         .into_iter()
         .flat_map(|typ| list_union_type(typ, field.is_list))
@@ -17,7 +17,7 @@ pub(crate) fn nested_create_input_field(ctx: &mut BuilderContext, field: &Relati
 }
 
 pub(crate) fn nested_connect_or_create_field(ctx: &mut BuilderContext, field: &RelationFieldRef) -> Option<InputField> {
-    create_input_objects::nested_connect_or_create_input_object(ctx, field).map(|input_object_type| {
+    create_objects::nested_connect_or_create_input_object(ctx, field).map(|input_object_type| {
         input_field(
             "connectOrCreate",
             list_union_object_type(input_object_type, field.is_list),
@@ -29,7 +29,7 @@ pub(crate) fn nested_connect_or_create_field(ctx: &mut BuilderContext, field: &R
 
 /// Builds "upsert" field for nested updates (on relation fields).
 pub(crate) fn nested_upsert_field(ctx: &mut BuilderContext, field: &RelationFieldRef) -> Option<InputField> {
-    update_input_objects::nested_upsert_input_object(ctx, field).map(|input_object_type| {
+    upsert_objects::nested_upsert_input_object(ctx, field).map(|input_object_type| {
         input_field("upsert", list_union_object_type(input_object_type, field.is_list), None).optional()
     })
 }
@@ -37,7 +37,7 @@ pub(crate) fn nested_upsert_field(ctx: &mut BuilderContext, field: &RelationFiel
 /// Builds "deleteMany" field for nested updates (on relation fields).
 pub(crate) fn nested_delete_many_field(ctx: &mut BuilderContext, field: &RelationFieldRef) -> Option<InputField> {
     if field.is_list {
-        let input_object = filter_input_objects::scalar_filter_object_type(ctx, &field.related_model());
+        let input_object = filter_objects::scalar_filter_object_type(ctx, &field.related_model());
         let input_type = InputType::object(input_object);
 
         Some(
@@ -55,7 +55,7 @@ pub(crate) fn nested_delete_many_field(ctx: &mut BuilderContext, field: &Relatio
 
 /// Builds "updateMany" field for nested updates (on relation fields).
 pub(crate) fn nested_update_many_field(ctx: &mut BuilderContext, field: &RelationFieldRef) -> Option<InputField> {
-    update_input_objects::nested_update_many_input_object(ctx, field).map(|input_object| {
+    update_many_objects::nested_update_many_input_object(ctx, field).map(|input_object| {
         let input_type = InputType::object(input_object);
 
         input_field(
@@ -105,7 +105,7 @@ pub(crate) fn nested_connect_input_field(ctx: &mut BuilderContext, field: &Relat
 }
 
 pub(crate) fn nested_update_input_field(ctx: &mut BuilderContext, field: &RelationFieldRef) -> InputField {
-    let input_object_type = update_input_objects::input_object_type_nested_update(ctx, field);
+    let input_object_type = update_one_objects::input_object_type_nested_update(ctx, field);
     input_field("update", list_union_object_type(input_object_type, field.is_list), None).optional()
 }
 
@@ -164,7 +164,7 @@ fn where_input_field<T>(ctx: &mut BuilderContext, name: T, field: &RelationField
 where
     T: Into<String>,
 {
-    let input_object_type = filter_input_objects::where_unique_object_type(ctx, &field.related_model());
+    let input_object_type = filter_objects::where_unique_object_type(ctx, &field.related_model());
     input_field(
         name.into(),
         list_union_object_type(input_object_type, field.is_list),
