+++ query-engine/core/src/schema_builder/input_types/objects/mod.rs
@@ -1,0 +1,9 @@
+pub(super) mod connect_or_create_objects;
+pub(super) mod create_objects;
+pub(super) mod filter_objects;
+pub(super) mod update_many_objects;
+pub(super) mod update_one_objects;
+pub(super) mod upsert_objects;
+
+use super::*;
+use prisma_models::{RelationFieldRef, ScalarFieldRef};
