+++ query-engine/core/src/schema_builder/output_types/mod.rs
@@ -3,3 +3,4 @@ pub(crate) mod output_objects;
 pub(crate) mod query_type;
 
 use super::*;
+use input_types::arguments;
